+++ src/script/event/EventRepository.test.ts
@@ -19,23 +19,14 @@
 
 import {BackendEvent, CONVERSATION_EVENT, USER_EVENT} from '@wireapp/api-client/lib/event/';
 
-import {Asset as ProtobufAsset} from '@wireapp/protocol-messaging';
-
-import {AssetTransferState} from 'src/script/assets/AssetTransferState';
-import {EventError} from 'src/script/error/EventError';
 import {ClientEvent} from 'src/script/event/Client';
 import {EventRepository} from 'src/script/event/EventRepository';
 import {NOTIFICATION_HANDLING_STATE} from 'src/script/event/NotificationHandlingState';
-import {createUuid} from 'Util/uuid';
 
-import {EventService} from './EventService';
 import {EventSource} from './EventSource';
 
 import {TestFactory} from '../../../test/helper/TestFactory';
 import {ClientConversationEvent} from '../conversation/EventBuilder';
-import {User} from '../entity/User';
-import {StatusType} from '../message/StatusType';
-import {EventRecord} from '../storage';
 
 const testFactory = new TestFactory();
 
@@ -46,64 +37,9 @@ describe('EventRepository', () => {
     return testFactory.exposeEventActors();
   });
 
-  describe('getCommonMessageUpdates', () => {
-    /** @see https://wearezeta.atlassian.net/browse/SQCORE-732 */
-    it('does not overwrite the seen status if a message gets edited', () => {
-      const originalEvent = {
-        category: 16,
-        conversation: 'a7f1187e-9396-44c9-8242-db9d3051dc89',
-        data: {
-          content: 'Original Text Which Has Been Seen By Someone Else',
-          expects_read_confirmation: true,
-          legal_hold_status: 1,
-          mentions: [],
-          previews: [],
-        },
-        from: '24de8432-03ba-439f-88f8-95bdc68b7bdd',
-        from_client_id: '79618bbe93e6821c',
-        id: 'c6269e58-fa82-4f6e-8264-263e09154871',
-        primary_key: '17',
-        read_receipts: [
-          {
-            time: '2021-06-10T19:47:19.570Z',
-            userId: 'b661e27f-24c6-4c52-a425-87a7b7f3df61',
-          },
-        ],
-        status: StatusType.SEEN,
-        time: '2021-06-10T19:47:16.071Z',
-        type: ClientEvent.CONVERSATION.MESSAGE_ADD,
-      } as any;
-
-      const editedEvent = {
-        conversation: 'a7f1187e-9396-44c9-8242-db9d3051dc89',
-        data: {
-          content: 'Edited Text Which Replaces The Original Text',
-          expects_read_confirmation: true,
-          mentions: [],
-          previews: [],
-          replacing_message_id: 'c6269e58-fa82-4f6e-8264-263e09154871',
-        },
-        from: '24de8432-03ba-439f-88f8-95bdc68b7bdd',
-        from_client_id: '79618bbe93e6821c',
-        id: 'caff044b-cb9c-47c6-833a-d4b76c678bcd',
-        status: StatusType.SENT,
-        time: '2021-06-10T19:47:23.706Z',
-        type: ClientEvent.CONVERSATION.MESSAGE_ADD,
-      } as any;
-
-      const updatedEvent = EventRepository['getCommonMessageUpdates'](originalEvent, editedEvent);
-      expect(updatedEvent.data.content).toBe('Edited Text Which Replaces The Original Text');
-      expect(updatedEvent.status).toBe(StatusType.SEEN);
-      expect(Object.keys(updatedEvent.read_receipts).length).toBe(1);
-    });
-  });
-
   describe('handleEvent', () => {
     beforeEach(() => {
       testFactory.event_repository!.notificationHandlingState(NOTIFICATION_HANDLING_STATE.WEB_SOCKET);
-      jest
-        .spyOn(testFactory.event_service!, 'saveEvent')
-        .mockReturnValue(Promise.resolve({data: 'dummy content'} as EventRecord));
       spyOn<any>(testFactory.event_repository!, 'distributeEvent');
     });
 
@@ -114,7 +50,6 @@ describe('EventRepository', () => {
           EventSource.NOTIFICATION_STREAM,
         )
         .then(() => {
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
           expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
         });
     });
@@ -126,7 +61,6 @@ describe('EventRepository', () => {
           EventSource.NOTIFICATION_STREAM,
         )
         .then(() => {
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
           expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
         });
     });
@@ -138,7 +72,6 @@ describe('EventRepository', () => {
           EventSource.NOTIFICATION_STREAM,
         )
         .then(() => {
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
           expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
         });
     });
@@ -154,7 +87,6 @@ describe('EventRepository', () => {
       } as BackendEvent;
 
       return testFactory.event_repository!['handleEvent']({event}, EventSource.NOTIFICATION_STREAM).then(() => {
-        expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
         expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
       });
     });
@@ -170,7 +102,6 @@ describe('EventRepository', () => {
       } as BackendEvent;
 
       return testFactory.event_repository!['handleEvent']({event}, EventSource.NOTIFICATION_STREAM).then(() => {
-        expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
         expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
       });
     });
@@ -186,18 +117,13 @@ describe('EventRepository', () => {
       } as BackendEvent;
 
       return testFactory.event_repository!['handleEvent']({event}, EventSource.NOTIFICATION_STREAM).then(() => {
-        expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
         expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
       });
     });
 
     it('accepts "conversation.voice-channel-deactivate" (missed call) events', async () => {
-      const eventServiceSpy = {
-        loadEvent: jest.fn().mockResolvedValue(undefined),
-        saveEvent: jest.fn().mockResolvedValue({data: 'dummy content'}),
-      } as unknown as EventService;
       const fakeProp: any = undefined;
-      const eventRepo = new EventRepository(eventServiceSpy, fakeProp, fakeProp, fakeProp);
+      const eventRepo = new EventRepository({} as any, fakeProp, fakeProp, fakeProp);
       eventRepo.notificationHandlingState(NOTIFICATION_HANDLING_STATE.WEB_SOCKET);
       jest.spyOn<any, any>(eventRepo, 'distributeEvent').mockImplementation(() => {});
 
@@ -212,7 +138,6 @@ describe('EventRepository', () => {
 
       await eventRepo['handleEvent']({event}, EventSource.NOTIFICATION_STREAM);
 
-      expect(eventServiceSpy.saveEvent).toHaveBeenCalled();
       expect(eventRepo['distributeEvent']).toHaveBeenCalled();
     });
 
@@ -228,383 +153,8 @@ describe('EventRepository', () => {
       } as ClientConversationEvent;
 
       return testFactory.event_repository.injectEvent(event).then(() => {
-        expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
         expect(testFactory.event_repository!['distributeEvent']).toHaveBeenCalled();
       });
-    });
-  });
-
-  describe('processEvent', () => {
-    let event: any;
-    let previously_stored_event: any;
-
-    beforeEach(() => {
-      event = {
-        conversation: createUuid(),
-        data: {
-          content: 'Lorem Ipsum',
-          previews: [],
-        },
-        from: createUuid(),
-        id: createUuid(),
-        time: new Date().toISOString(),
-        type: ClientEvent.CONVERSATION.MESSAGE_ADD,
-      };
-
-      jest
-        .spyOn(testFactory.event_service!, 'saveEvent')
-        .mockImplementation(saved_event => Promise.resolve(saved_event as any));
-    });
-
-    it('saves an event with a previously not used ID', () => {
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockClear();
-
-      return testFactory.event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM).then(() => {
-        expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
-      });
-    });
-
-    it('ignores an event with an ID previously used by another user', () => {
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      previously_stored_event.from = createUuid();
-      jest
-        .spyOn(testFactory.event_service!, 'loadEvent')
-        .mockImplementation(() => Promise.resolve(previously_stored_event));
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('ignores a non-"text message" with an ID previously used by the same user', () => {
-      event.type = ClientEvent.CALL.E_CALL;
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      jest
-        .spyOn(testFactory.event_service!, 'loadEvent')
-        .mockImplementation(() => Promise.resolve(previously_stored_event));
-
-      return testFactory
-        .event_repository!['handleEventSaving'](event)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('ignores a plain text message with an ID previously used by the same user for a non-"text message"', () => {
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      previously_stored_event.type = ClientEvent.CALL.E_CALL;
-      jest
-        .spyOn(testFactory.event_service!, 'loadEvent')
-        .mockImplementation(() => Promise.resolve(previously_stored_event));
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('ignores a plain text message with an ID previously used by the same user', () => {
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      jest
-        .spyOn(testFactory.event_service!, 'loadEvent')
-        .mockImplementation(() => Promise.resolve(previously_stored_event));
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('ignores a text message with link preview with an ID previously used by the same user for a text message with link preview', () => {
-      event.data.previews.push(1);
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(previously_stored_event as any);
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('ignores a text message with link preview with an ID previously used by the same user for a text message different content', () => {
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(previously_stored_event as any);
-
-      event.data.previews.push(1);
-      event.data.content = 'Ipsum loren';
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then(() => fail('Method should have thrown an error'))
-        .catch(error => {
-          expect(error).toEqual(jasmine.any(EventError));
-          expect(error.type).toBe(EventError.TYPE.VALIDATION_FAILED);
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('saves a text message with link preview with an ID previously used by the same user for a plain text message', () => {
-      previously_stored_event = JSON.parse(JSON.stringify(event));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(previously_stored_event);
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation(() => Promise.resolve(previously_stored_event));
-
-      const initial_time = event.time;
-      const changed_time = new Date(new Date(event.time).getTime() + 60 * 1000).toISOString();
-      event.data.previews.push(1);
-      event.time = changed_time;
-
-      return testFactory
-        .event_repository!['processEvent'](event, EventSource.NOTIFICATION_STREAM)
-        .then((saved_event: any) => {
-          expect(saved_event.time).toEqual(initial_time);
-          expect(saved_event.time).not.toEqual(changed_time);
-          expect(saved_event.primary_key).toEqual(previously_stored_event.primary_key);
-          expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-        });
-    });
-
-    it('ignores edit message with missing associated original message', () => {
-      const linkPreviewEvent = JSON.parse(JSON.stringify(event));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue({} as any);
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation(() => Promise.resolve({} as EventRecord));
-
-      linkPreviewEvent.data.replacing_message_id = 'initial_message_id';
-
-      return testFactory
-        .event_repository!['handleEventSaving'](linkPreviewEvent)
-        .then(() => fail('Should have thrown an error'))
-        .catch(() => {
-          expect(testFactory.event_service!.replaceEvent).not.toHaveBeenCalled();
-          expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        });
-    });
-
-    it('updates edited messages when link preview arrives', () => {
-      const replacingId = 'old-replaced-message-id';
-      const storedEvent = {
-        ...event,
-        data: {...event.data, replacing_message_id: replacingId},
-      } as EventRecord;
-      const linkPreviewEvent = {...event};
-      jest
-        .spyOn(testFactory.event_service!, 'loadEvent')
-        .mockImplementation((conversationId: string, messageId: string) => {
-          return messageId === replacingId ? Promise.resolve(undefined) : Promise.resolve(storedEvent as any);
-        });
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation((ev: EventRecord) => Promise.resolve(ev));
-
-      linkPreviewEvent.data.replacing_message_id = replacingId;
-      linkPreviewEvent.data.previews = ['preview'];
-
-      return testFactory.event_repository!['handleEventSaving'](linkPreviewEvent).then((updatedEvent: any) => {
-        expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-        expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        expect(updatedEvent?.data.previews[0]).toEqual('preview');
-      });
-    });
-
-    it('updates edited messages', () => {
-      const originalMessage = JSON.parse(JSON.stringify(event));
-      originalMessage.reactions = ['user-id'];
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(originalMessage as any);
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation((updates: EventRecord) => Promise.resolve(updates));
-
-      const initial_time = event.time;
-      const changed_time = new Date(new Date(event.time).getTime() + 60 * 1000).toISOString();
-      originalMessage.primary_key = 12;
-      event.id = createUuid();
-      event.data.content = 'new content';
-      event.data.replacing_message_id = originalMessage.id;
-      event.time = changed_time;
-
-      return testFactory.event_repository!['handleEventSaving'](event).then((updatedEvent: any) => {
-        expect(updatedEvent.time).toEqual(initial_time);
-        expect(updatedEvent.time).not.toEqual(changed_time);
-        expect(updatedEvent.data.content).toEqual('new content');
-        expect(updatedEvent.primary_key).toEqual(originalMessage.primary_key);
-        expect(Object.keys(updatedEvent.reactions).length).toEqual(0);
-        expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-      });
-    });
-
-    it('updates link preview when edited', () => {
-      const replacingId = 'replaced-message-id';
-      const storedEvent = {
-        ...event,
-        data: {...event.data, previews: ['preview']},
-      } as any;
-      const editEvent = {...event} as any;
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(storedEvent as any);
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation((ev: EventRecord) => Promise.resolve(ev));
-
-      editEvent.data.replacing_message_id = replacingId;
-
-      return testFactory.event_repository!['handleEventSaving'](editEvent).then((updatedEvent: any) => {
-        expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-        expect(testFactory.event_service!.saveEvent).not.toHaveBeenCalled();
-        expect(updatedEvent?.data.previews.length).toEqual(0);
-      });
-    });
-
-    it('saves a conversation.asset-add event', () => {
-      const assetAddEvent = {...event, type: ClientEvent.CONVERSATION.ASSET_ADD};
-
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockClear();
-
-      return testFactory
-        .event_repository!['processEvent'](assetAddEvent, EventSource.NOTIFICATION_STREAM)
-        .then(updatedEvent => {
-          expect(updatedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-          expect(testFactory.event_service!.saveEvent).toHaveBeenCalled();
-        });
-    });
-
-    it('deletes cancelled conversation.asset-add event', async () => {
-      const fromIds = [
-        // cancel from an other user
-        createUuid(),
-        // cancel from the self user
-        testFactory.user_repository['userState'].self().id,
-      ];
-
-      const loadEventSpy = jest.spyOn(testFactory.event_service!, 'loadEvent');
-      const deleteEventSpy = jest.spyOn(testFactory.event_service!, 'deleteEvent');
-      for (const fromId of fromIds) {
-        const assetAddEvent = {...event, from: fromId, type: ClientEvent.CONVERSATION.ASSET_ADD};
-        const assetCancelEvent = {
-          ...assetAddEvent,
-          data: {reason: ProtobufAsset.NotUploaded.CANCELLED, status: AssetTransferState.UPLOAD_FAILED},
-          time: '2017-09-06T09:43:36.528Z',
-        };
-
-        loadEventSpy.mockResolvedValue(assetAddEvent as any);
-        deleteEventSpy.mockImplementation(() => Promise.resolve(1));
-
-        const savedEvent = await testFactory.event_repository!['processEvent'](
-          assetCancelEvent,
-          EventSource.NOTIFICATION_STREAM,
-        );
-        expect(savedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-        expect(testFactory.event_service!.deleteEvent).toHaveBeenCalled();
-      }
-    });
-
-    it('deletes other user failed upload for conversation.asset-add event', () => {
-      const assetAddEvent = {...event, type: ClientEvent.CONVERSATION.ASSET_ADD};
-      const assetUploadFailedEvent = {
-        ...assetAddEvent,
-        data: {reason: ProtobufAsset.NotUploaded.FAILED, status: AssetTransferState.UPLOAD_FAILED},
-        time: '2017-09-06T09:43:36.528Z',
-      };
-
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(assetAddEvent as any);
-      jest.spyOn(testFactory.event_service!, 'deleteEvent').mockImplementation(() => Promise.resolve(1));
-
-      return testFactory
-        .event_repository!['processEvent'](assetUploadFailedEvent, EventSource.NOTIFICATION_STREAM)
-        .then(savedEvent => {
-          expect(savedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-          expect(testFactory.event_service!.deleteEvent).toHaveBeenCalled();
-        });
-    });
-
-    it('updates self failed upload for conversation.asset-add event', async () => {
-      const assetAddEvent: EventRecord = {...event, type: ClientEvent.CONVERSATION.ASSET_ADD};
-      const assetUploadFailedEvent = {
-        ...assetAddEvent,
-        data: {reason: ProtobufAsset.NotUploaded.FAILED, status: AssetTransferState.UPLOAD_FAILED},
-        time: '2017-09-06T09:43:36.528Z',
-      } as any;
-
-      jest
-        .spyOn(testFactory.event_repository!['userState'], 'self')
-        .mockReturnValue(new User(assetAddEvent.from) as any);
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(assetAddEvent as any);
-      jest
-        .spyOn(testFactory.event_service!, 'updateEventAsUploadFailed')
-        .mockImplementation(() => Promise.resolve(assetUploadFailedEvent));
-
-      const savedEvent = await testFactory.event_repository!['processEvent'](
-        assetUploadFailedEvent,
-        EventSource.NOTIFICATION_STREAM,
-      );
-      expect(savedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-      expect(testFactory.event_service!.updateEventAsUploadFailed).toHaveBeenCalled();
-    });
-
-    it('handles conversation.asset-add state update event', () => {
-      const initialAssetEvent = {...event, type: ClientEvent.CONVERSATION.ASSET_ADD};
-
-      const updateStatusEvent = {
-        ...initialAssetEvent,
-        data: {status: AssetTransferState.UPLOADED},
-        time: '2017-09-06T09:43:36.528Z',
-      };
-
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation(eventToUpdate => Promise.resolve(eventToUpdate));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(initialAssetEvent as any);
-
-      return testFactory
-        .event_repository!['processEvent'](updateStatusEvent, EventSource.NOTIFICATION_STREAM)
-        .then((updatedEvent: any) => {
-          expect(updatedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-          expect(updatedEvent.data.status).toEqual(updateStatusEvent.data.status);
-          expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-        });
-    });
-
-    it('updates video when preview is received', () => {
-      const initialAssetEvent = {...event, type: ClientEvent.CONVERSATION.ASSET_ADD};
-
-      const AssetPreviewEvent = {
-        ...initialAssetEvent,
-        data: {status: AssetTransferState.UPLOADED},
-        time: '2017-09-06T09:43:36.528Z',
-      };
-
-      jest
-        .spyOn(testFactory.event_service!, 'replaceEvent')
-        .mockImplementation(eventToUpdate => Promise.resolve(eventToUpdate));
-      jest.spyOn(testFactory.event_service!, 'loadEvent').mockResolvedValue(initialAssetEvent as any);
-
-      return testFactory
-        .event_repository!['processEvent'](AssetPreviewEvent, EventSource.NOTIFICATION_STREAM)
-        .then((updatedEvent: EventRecord) => {
-          expect(updatedEvent.type).toEqual(ClientEvent.CONVERSATION.ASSET_ADD);
-          expect(testFactory.event_service!.replaceEvent).toHaveBeenCalled();
-        });
     });
   });
 });
