+++ src/script/event/EventRepository.ts
@@ -29,32 +29,26 @@ import ko from 'knockout';
 import {container} from 'tsyringe';
 
 import {Account, ConnectionState, ProcessedEventPayload} from '@wireapp/core';
-import {Asset as ProtobufAsset} from '@wireapp/protocol-messaging';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {getLogger, Logger} from 'Util/Logger';
 import {queue} from 'Util/PromiseQueue';
 import {TIME_IN_MILLIS} from 'Util/TimeUtil';
 
-import {ClientEvent, CONVERSATION} from './Client';
+import {ClientEvent} from './Client';
 import {EventMiddleware, EventProcessor, IncomingEvent} from './EventProcessor';
 import type {EventService} from './EventService';
 import {EventSource} from './EventSource';
 import {EVENT_TYPE} from './EventType';
-import {EventTypeHandling} from './EventTypeHandling';
 import {EventValidation} from './EventValidation';
 import {validateEvent} from './EventValidator';
 import {NOTIFICATION_HANDLING_STATE} from './NotificationHandlingState';
 import type {NotificationService} from './NotificationService';
 
-import {AssetTransferState} from '../assets/AssetTransferState';
-import {AssetAddEvent, ClientConversationEvent, EventBuilder, MessageAddEvent} from '../conversation/EventBuilder';
+import {ClientConversationEvent, EventBuilder} from '../conversation/EventBuilder';
 import {CryptographyMapper} from '../cryptography/CryptographyMapper';
 import {CryptographyError} from '../error/CryptographyError';
 import {EventError} from '../error/EventError';
-import {categoryFromEvent} from '../message/MessageCategorization';
-import {isEventRecordFailed, isEventRecordWithFederationError} from '../message/StatusType';
-import type {EventRecord, StoredEvent} from '../storage';
 import type {ServerTimeHandler} from '../time/serverTimeHandler';
 import {EventName} from '../tracking/EventName';
 import {UserState} from '../user/UserState';
@@ -430,12 +424,6 @@ export class EventRepository {
       event = await eventProcessMiddleware.processEvent(event);
     }
 
-    const shouldSaveEvent = EventTypeHandling.STORE.includes(event.type as CONVERSATION_EVENT);
-    if (shouldSaveEvent) {
-      const savedEvent = await this.handleEventSaving(event);
-      event = savedEvent ?? event;
-    }
-
     return this.handleEventDistribution(event, source);
   }
 
@@ -462,212 +450,5 @@ export class EventRepository {
       }
     }
     return this.distributeEvent(event, source);
-  }
-
-  /**
-   * Handle a mapped event, check for malicious ID use and save it.
-   *
-   * @param event Backend event extracted from notification stream
-   * @returns Resolves with the saved event
-   */
-  private async handleEventSaving(event: IncomingEvent) {
-    const conversationId = 'conversation' in event && event.conversation;
-
-    // first check if a message that should be replaced exists in DB
-    if (event.type === ClientEvent.CONVERSATION.MESSAGE_ADD) {
-      const mappedData = event.data;
-      const eventToReplace = mappedData.replacing_message_id
-        ? await this.eventService.loadEvent(conversationId, mappedData.replacing_message_id)
-        : undefined;
-
-      const hasLinkPreview = mappedData.previews && mappedData.previews.length;
-      const isReplacementWithoutOriginal = !eventToReplace && mappedData.replacing_message_id;
-      if (isReplacementWithoutOriginal && !hasLinkPreview) {
-        // the only valid case of a replacement with no original message is when an edited message gets a link preview
-        this.throwValidationError(event, 'Edit event without original event');
-      }
-
-      if (eventToReplace?.type === CONVERSATION.MESSAGE_ADD) {
-        return this.handleEventReplacement(eventToReplace, event);
-      }
-    }
-
-    // check for duplicates (same id)
-    const storedEvent = 'id' in event ? await this.eventService.loadEvent(conversationId, event.id) : undefined;
-
-    return storedEvent
-      ? this.handleDuplicatedEvent(storedEvent, event)
-      : this.eventService.saveEvent(event as EventRecord);
-  }
-
-  private handleEventReplacement(originalEvent: StoredEvent<MessageAddEvent>, newEvent: MessageAddEvent) {
-    if (originalEvent.from !== newEvent.from) {
-      const logMessage = `ID previously used by user '${newEvent.from}'`;
-      const errorMessage = 'ID reused by other user';
-      this.throwValidationError(newEvent, errorMessage, logMessage);
-    }
-    const newData = newEvent.data;
-    const primaryKeyUpdate = {primary_key: originalEvent.primary_key};
-    const isLinkPreviewEdit = newData?.previews && !!newData?.previews.length;
-
-    const commonUpdates = EventRepository.getCommonMessageUpdates(originalEvent, newEvent);
-
-    const specificUpdates = isLinkPreviewEdit
-      ? this.getUpdatesForMessage(originalEvent, newEvent)
-      : EventRepository.getUpdatesForEditMessage(originalEvent, newEvent);
-
-    const updates = {...specificUpdates, ...commonUpdates};
-
-    const identifiedUpdates = {...primaryKeyUpdate, ...updates};
-    return this.eventService.replaceEvent(identifiedUpdates);
-  }
-
-  private handleDuplicatedEvent(originalEvent: EventRecord, newEvent: IncomingEvent) {
-    switch (newEvent.type) {
-      case ClientEvent.CONVERSATION.ASSET_ADD:
-        return this.handleAssetUpdate(originalEvent, newEvent);
-
-      case ClientEvent.CONVERSATION.MESSAGE_ADD:
-        return this.handleMessageUpdate(originalEvent, newEvent);
-
-      default:
-        this.throwValidationError(newEvent, `Forbidden type '${newEvent.type}' for duplicate events`);
-    }
-  }
-
-  private async handleAssetUpdate(originalEvent: EventRecord, newEvent: AssetAddEvent) {
-    if (originalEvent.type !== ClientEvent.CONVERSATION.ASSET_ADD) {
-      this.throwValidationError(newEvent, 'Trying to update a non-asset message as an asset message');
-    }
-    const newEventData = newEvent.data;
-    // the preview status is not sent by the client so we fake a 'preview' status in order to cleanly handle it in the switch statement
-    const ASSET_PREVIEW = 'preview';
-    // similarly, no status is sent by the client when we retry sending a failed message
-    const RETRY_EVENT = 'retry';
-    const isPreviewEvent = !newEventData.status && !!newEventData.preview_key;
-    const isRetryEvent = !!newEventData.content_length;
-    const handledEvent = isRetryEvent ? RETRY_EVENT : newEventData.status;
-    const previewStatus = isPreviewEvent ? ASSET_PREVIEW : handledEvent;
-
-    const updateEvent = () => {
-      const updatedData = {...originalEvent.data, ...newEventData};
-      const updatedEvent = {...originalEvent, data: updatedData};
-      return this.eventService.replaceEvent(updatedEvent);
-    };
-
-    switch (previewStatus) {
-      case ASSET_PREVIEW:
-      case RETRY_EVENT:
-      case AssetTransferState.UPLOADED: {
-        return updateEvent();
-      }
-
-      case AssetTransferState.UPLOAD_FAILED: {
-        // case of both failed or canceled upload
-        const fromOther = newEvent.from !== this.userState.self().id;
-        const sameSender = newEvent.from === originalEvent.from;
-        const selfCancel = !fromOther && newEvent.data.reason === ProtobufAsset.NotUploaded.CANCELLED;
-        // we want to delete the event in the case of an error from the remote client or a cancel on the user's own client
-        const shouldDeleteEvent = (fromOther || selfCancel) && sameSender;
-        if (shouldDeleteEvent) {
-          await this.eventService.deleteEvent(newEvent.conversation, newEvent.id);
-          return newEvent;
-        }
-        return this.eventService.updateEventAsUploadFailed(originalEvent.primary_key, newEvent.data.reason);
-      }
-
-      default: {
-        this.throwValidationError(newEvent, `Unhandled asset status update '${newEvent.data.status}'`);
-      }
-    }
-  }
-
-  private handleMessageUpdate(originalEvent: EventRecord, newEvent: MessageAddEvent) {
-    const newEventData = newEvent.data;
-    const originalData = originalEvent.data;
-
-    if (originalEvent.from !== newEvent.from) {
-      const logMessage = `ID previously used by user '${newEvent.from}'`;
-      const errorMessage = 'ID reused by other user';
-      return this.throwValidationError(newEvent, errorMessage, logMessage);
-    }
-
-    const containsLinkPreview = newEventData.previews && !!newEventData.previews.length;
-    const isRetryAttempt = isEventRecordFailed(originalEvent) || isEventRecordWithFederationError(originalEvent);
-
-    if (!containsLinkPreview && !isRetryAttempt) {
-      const errorMessage =
-        'Message duplication event invalid: original message did not fail to send and does not contain link preview';
-      return this.throwValidationError(newEvent, errorMessage);
-    }
-
-    const textContentMatches = newEventData.content === (originalData as any).content;
-    if (!textContentMatches) {
-      const errorMessage = 'ID of link preview reused';
-      const logMessage = 'Text content for message duplication not matching';
-      return this.throwValidationError(newEvent, errorMessage, logMessage);
-    }
-
-    const bothAreMessageAddType = newEvent.type === originalEvent.type;
-    if (!bothAreMessageAddType) {
-      return this.throwValidationError(newEvent, 'ID reused by same user');
-    }
-
-    const updates = this.getUpdatesForMessage(originalEvent, newEvent);
-    const identifiedUpdates = {primary_key: originalEvent.primary_key, ...updates};
-    return this.eventService.replaceEvent(identifiedUpdates);
-  }
-
-  private static getCommonMessageUpdates(originalEvent: StoredEvent<MessageAddEvent>, newEvent: MessageAddEvent) {
-    return {
-      ...newEvent,
-      data: {...newEvent.data, expects_read_confirmation: originalEvent.data.expects_read_confirmation},
-      edited_time: newEvent.time,
-      read_receipts: !newEvent.read_receipts ? originalEvent.read_receipts : newEvent.read_receipts,
-      status: !newEvent.status || newEvent.status < originalEvent.status ? originalEvent.status : newEvent.status,
-      time: originalEvent.time,
-      version: 1,
-    };
-  }
-
-  private static getUpdatesForEditMessage(originalEvent: EventRecord, newEvent: MessageAddEvent): MessageAddEvent {
-    // Remove reactions, so that likes (hearts) don't stay when a message's text gets edited
-    return {...newEvent, reactions: {}};
-  }
-
-  private getUpdatesForMessage(originalEvent: EventRecord, newEvent: MessageAddEvent) {
-    const newData = newEvent.data;
-    const originalData = originalEvent.data;
-    const updatingLinkPreview = !!(originalData as any).previews.length;
-    if (updatingLinkPreview) {
-      this.throwValidationError(newEvent, 'ID of link preview reused');
-    }
-
-    const textContentMatches = !newData.previews?.length || newData.content === (originalData as any).content;
-    if (!textContentMatches) {
-      const logMessage = 'Text content for message duplication not matching';
-      const errorMessage = 'ID of duplicated message reused';
-      this.throwValidationError(newEvent, errorMessage, logMessage);
-    }
-
-    return {
-      ...newEvent,
-      category: categoryFromEvent(newEvent),
-      ephemeral_expires: originalEvent.ephemeral_expires,
-      ephemeral_started: originalEvent.ephemeral_started,
-      ephemeral_time: originalEvent.ephemeral_time,
-      server_time: newEvent.time,
-      time: originalEvent.time,
-      version: originalEvent.version,
-    };
-  }
-
-  private throwValidationError(event: IncomingEvent, errorMessage: string, logMessage?: string): never {
-    const conversation = 'conversation' in event && event.conversation;
-    const from = 'from' in event && event.from;
-
-    const baseLogMessage = `Ignored '${event.type}' in '${conversation}' from '${from}''`;
-    this.logger.warn(`${baseLogMessage} ${logMessage || errorMessage}`);
-    throw new EventError(EventError.TYPE.VALIDATION_FAILED, `Event validation failed: ${errorMessage}`);
   }
 }
